requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
scipy
statsmodels
pulp
plotly
streamlit
yfinance
openpyxl
Pillow
